refactor: UserRepository using CacheUser
use decorating patron in CacheUser in this way we can have cache and use it whenever we want.

UserController

Resolve: $19
